PR리뷰 반영 - Test를 위한 코드를 test package 로 이동 https://github.com/next-step/kotlin-minesweeper/pull/334#discussion_r1257876839
