+++ src/main/kotlin/domain/MineSweeperBoardGenerator.kt
@@ -31,30 +31,3 @@ abstract class MineSweeperBoardGenerator(val boardSize: BoardSize, private val m
             }
     }
 }
-
-class RandomMineSweeperBoardGenerator(boardSize: BoardSize, mines: Int) :
-    MineSweeperBoardGenerator(boardSize, mines)
-
-class TestMineSweeperBoardGenerator : MineSweeperBoardGenerator(BoardSize(10, 10), 10) {
-    override fun generate(): MineSweeperBoard {
-        val mineSweeperBoard = MineSweeperBoard(boardSize)
-        mineSweeperBoard.putMines(
-            Position(3, 0),
-            Position(7, 0),
-
-            Position(2, 1),
-            Position(4, 1),
-
-            Position(0, 4),
-
-            Position(6, 5),
-
-            Position(2, 6),
-            Position(6, 6),
-
-            Position(6, 7),
-            Position(9, 7),
-        )
-        return mineSweeperBoard
-    }
-}
