+++ src/main/kotlin/domain/RandomMineSweeperBoardGenerator.kt
@@ -1,0 +1,4 @@
+package domain
+
+class RandomMineSweeperBoardGenerator(boardSize: BoardSize, mines: Int) :
+    MineSweeperBoardGenerator(boardSize, mines)
